Update / fix: Allow use of generic models by avoiding problem of Self type constraints
Details:

Remove Equatable inheritance from IdentifiableType to allow for generics to to be used as models, avoiding the problem with Generics and conformance to types with Self requirements.  This update allows the best of both approaches - concrete types can conform to Equatable, and generics can be used by providing an explicit comparison function.  Meanwhile, the POP technique of a default Equatable implementation for IdentifiableType still exists, so that any model that implements both, gets the default implementation of Equatable, as before.